Refactor ConfirmMode to config package
