+++ src/test/java/org/mvplugins/multiverse/core/commands/RegenCommandTest.kt
@@ -2,7 +2,7 @@ package org.mvplugins.multiverse.core.commands
 
 import org.bukkit.Bukkit
 import org.bukkit.GameRule
-import org.mvplugins.multiverse.core.commandtools.ConfirmMode
+import org.mvplugins.multiverse.core.config.ConfirmMode
 import org.mvplugins.multiverse.core.config.MVCoreConfig
 import org.mvplugins.multiverse.core.world.LoadedMultiverseWorld
 import org.mvplugins.multiverse.core.world.options.CreateWorldOptions
